getTransformFunction - check for exact matches before looking for regex matches
